Moved https get request back to main thread
