+++ app/main/ipcHandlers.ts
@@ -1,3 +1,4 @@
+import axios from 'axios';
 import { app, shell, BrowserWindow, ipcMain, dialog } from 'electron';
 import { PrintErrorTypes } from '~/utils/types';
 import ipcCommands from '~/constants/ipcCommands.json';
@@ -25,6 +26,42 @@ async function print(body: string, printWindow: BrowserWindow) {
     });
 }
 
+async function httpsGet(
+    urlString: string,
+    params: Record<string, string>
+): Promise<string> {
+    // Setup timeout for axios (it's a little weird, as default timeout
+    // settings in axios only concern themselves with response timeout,
+    // not a connect timeout).
+    const source = axios.CancelToken.source();
+    const timeout = setTimeout(() => {
+        source.cancel();
+    }, 60000);
+
+    const searchParams = new URLSearchParams(params);
+    let urlGet: string;
+    if (Object.entries(params).length === 0) {
+        urlGet = urlString;
+    } else {
+        urlGet = `${urlString}?${searchParams.toString()}`;
+    }
+
+    const response = await axios.get(urlGet, {
+        cancelToken: source.token,
+        maxRedirects: 0,
+        // We also want to accept a 302 redirect, as that is used by the
+        // identity provider flow
+        validateStatus: (status: number) => status >= 200 && status <= 302,
+    });
+    clearTimeout(timeout);
+
+    return JSON.stringify({
+        data: response.data,
+        headers: response.headers,
+        status: response.status,
+    });
+}
+
 export default function initializeIpcHandlers(printWindow: BrowserWindow) {
     // Returns the path to userdata.
     ipcMain.handle(ipcCommands.getUserDataPath, async () => {
@@ -39,6 +76,13 @@ export default function initializeIpcHandlers(printWindow: BrowserWindow) {
     ipcMain.handle(ipcCommands.openUrl, (_event, url: string) => {
         shell.openExternal(url);
     });
+
+    ipcMain.handle(
+        ipcCommands.httpsGet,
+        (_event, url: string, params: Record<string, string>) => {
+            return httpsGet(url, params);
+        }
+    );
 
     // Provides access to save file dialog from renderer processes.
     ipcMain.handle(ipcCommands.saveFileDialog, async (_event, opts) => {
